Bibdk2021 1170 no hitcount (#501)
* in progress - remove hitcount from desktop and mobile

* in progress - remove hitcount from desktop and mobile

* in progress - filtebutton + no hitcount

* removed count

* remove unused css

* remove spacing on result section

* fixed some tests
