+++ src/components/base/translate/Translate.json
@@ -1138,7 +1138,7 @@
         "en": "Sort alphabetically"
       },
       "label-sortorder-alphabetically": {
-        "da": "Sorter efter antal",
+        "da": "Sorter efter hyppighed",
         "en": "Sort by numbers"
       },
       "label-canAlwaysBeLoaned": {
